[ui] Add Cart component showing all selected statistical units
