+++ app/src/app/search/components/search-result-table.tsx
@@ -1,22 +1,14 @@
-import {Table, TableBody, TableCell, TableHead, TableHeader, TableRow} from "@/components/ui/table";
+import {Table, TableBody, TableCell, TableRow} from "@/components/ui/table";
 import {useSearchContext} from "../search-provider";
-import SortableTableHead from "@/app/search/components/sortable-table-head";
-import {SearchResultTableRow} from "@/app/search/components/search-result-table-row";
+import {StatisticalUnitTableRow} from "@/app/search/components/statistical-unit-table-row";
+import {StatisticalUnitTableHeader} from "@/app/search/components/statistical-unit-table-header";
 
 export default function SearchResultTable() {
     const {searchResult} = useSearchContext();
 
     return (
         <Table>
-            <TableHeader className="bg-gray-100">
-                <TableRow>
-                    <SortableTableHead name="name">Name</SortableTableHead>
-                    <SortableTableHead className="text-left" name="physical_region_path">Region</SortableTableHead>
-                    <SortableTableHead className="text-right" name="employees">Employees</SortableTableHead>
-                    <SortableTableHead className="text-left" name="primary_activity_category_path">Activity Category</SortableTableHead>
-                    <TableHead />
-                </TableRow>
-            </TableHeader>
+            <StatisticalUnitTableHeader/>
             <TableBody>
                 {
                     !searchResult?.statisticalUnits.length && (
@@ -28,7 +20,10 @@ export default function SearchResultTable() {
                 {
                     searchResult?.statisticalUnits.map((unit) => {
                             return (
-                                <SearchResultTableRow key={`${unit.unit_id}-${unit.unit_type}`} unit={unit}/>
+                                <StatisticalUnitTableRow
+                                    key={`${unit.unit_id}-${unit.unit_type}`}
+                                    unit={unit}
+                                />
                             )
                         }
                     )
